requirements.txt
tensorflow==2.13.0
opencv-python==4.8.0.76
numpy==1.24.3
matplotlib==3.7.2
pillow==10.0.0
streamlit==1.28.0
requests==2.31.0